+++ package.json
@@ -22,6 +22,7 @@
   },
   "homepage": "https://github.com/FictionalDevelopers/expenses-tracker#readme",
   "dependencies": {
+    "@material-ui/core": "^4.4.3",
     "concurrently": "^4.1.2",
     "connected-react-router": "^6.5.2",
     "dotenv": "^8.1.0",
