+++ src/App.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import { Switch, Route, BrowserRouter } from "react-router-dom";
+import { Switch, BrowserRouter } from "react-router-dom";
 import Root from './routes/Root';
 import Login from './routes/Login';
-import PrivateRoute from "./common/PrivatRoute";
-import PublicOnlyRoute from "./common/PublicOnlyRoute";
+import PrivateRoute from "./common/components/PrivatRoute";
+import PublicOnlyRoute from "./common/components/PublicOnlyRoute";
 
 const App = () => {
     return (
